pilot-core/libqrot/libquat: merge libqrot into libquat and adjust pilot-core
 - libqrot is now part of libquat
 - libquat is now dependant on libvec3d
 - changes across pilot-core to fit new libraries structure

JIRA: PP-98
