+++ ekf/kalman_update_imu.c
@@ -24,7 +24,6 @@
 
 #include <vec.h>
 #include <quat.h>
-#include <qrot.h>
 #include <matrix.h>
 
 #include "kalman_implem.h"
