+++ quadcontrol/Makefile
@@ -6,5 +6,5 @@
 
 NAME := quad-control
 LOCAL_SRCS := control.c pid.c mma.c
-DEP_LIBS := libekf libmatrix libvec3d libquat libqrot
+DEP_LIBS := libekf libmatrix libvec3d libquat
 include $(binary.mk)
